Fixes based on PR comments
