+++ services/api-server/src/simcore_service_api_server/api/dependencies/models_schemas_function_filters.py
@@ -1,20 +1,9 @@
-from typing import Annotated, Any
+from typing import Annotated
 
 from fastapi import Query
 from models_library.functions import FunctionJobCollectionsListFilters
-from pydantic.fields import FieldInfo
 
-
-def _get_query_params(field: FieldInfo) -> dict[str, Any]:
-    params = {}
-
-    if field.description:
-        params["description"] = field.description
-    if field.examples:
-        params["example"] = next(
-            (example for example in field.examples if "*" in example), field.examples[0]
-        )
-    return params
+from ._utils import _get_query_params
 
 
 def get_function_job_collections_filters(
